+++ libr/include/r_util/r_strs.h
@@ -41,55 +41,7 @@ static inline bool r_strs_empty(RStrs s) {
 	return !s.a || s.b <= s.a;
 }
 
-static inline char r_strs_at(RStrs s, size_t i) {
-	return (i < r_strs_len (s))? s.a[i]: 0;
-}
-
-static inline char r_strs_first(RStrs s) {
-	return r_strs_empty (s)? 0: s.a[0];
-}
-
-static inline char r_strs_last(RStrs s) {
-	return r_strs_empty (s)? 0: s.b[-1];
-}
-
 /* Subslicing */
-static inline RStrs r_strs_head(RStrs s, size_t n) {
-	const size_t l = r_strs_len (s);
-	if (n > l) {
-		n = l;
-	}
-	RStrs r = { s.a, s.a + n };
-	return r;
-}
-
-static inline RStrs r_strs_tail(RStrs s, size_t n) {
-	const size_t l = r_strs_len (s);
-	if (n > l) {
-		n = l;
-	}
-	RStrs r = { s.b - n, s.b };
-	return r;
-}
-
-static inline RStrs r_strs_drop(RStrs s, size_t n) {
-	const size_t l = r_strs_len (s);
-	if (n > l) {
-		n = l;
-	}
-	RStrs r = { s.a + n, s.b };
-	return r;
-}
-
-static inline RStrs r_strs_drop_back(RStrs s, size_t n) {
-	const size_t l = r_strs_len (s);
-	if (n > l) {
-		n = l;
-	}
-	RStrs r = { s.a, s.b - n };
-	return r;
-}
-
 static inline RStrs r_strs_sub(RStrs s, size_t from, size_t to) {
 	const size_t l = r_strs_len (s);
 	if (from > l) {
@@ -115,42 +67,17 @@ static inline bool r_strs_advance(RStrs *s, size_t n) {
 	return true;
 }
 
-static inline char r_strs_pop_front(RStrs *s) {
-	if (r_strs_empty (*s)) {
-		return 0;
-	}
-	const char c = *s->a;
-	s->a++;
-	return c;
-}
-
-static inline char r_strs_pop_back(RStrs *s) {
-	if (r_strs_empty (*s)) {
-		return 0;
-	}
-	s->b--;
-	return *s->b;
-}
-
 /* Whitespace trimming */
-static inline void r_strs_ltrim(RStrs *s) {
+static inline void r_strs_trim(RStrs *s) {
 	while (s->a < s->b && isspace ((unsigned char)*s->a)) {
 		s->a++;
 	}
-}
-
-static inline void r_strs_rtrim(RStrs *s) {
 	while (s->b > s->a && isspace ((unsigned char)s->b[-1])) {
 		s->b--;
 	}
 }
 
-static inline void r_strs_trim(RStrs *s) {
-	r_strs_ltrim (s);
-	r_strs_rtrim (s);
-}
-
-/* Equality / prefix / suffix — canonical slice versions */
+/* Equality / prefix — canonical slice versions */
 static inline bool r_strs_equals(RStrs a, RStrs b) {
 	const size_t la = r_strs_len (a);
 	return la == r_strs_len (b) && (la == 0 || !memcmp (a.a, b.a, la));
@@ -159,12 +86,6 @@ static inline bool r_strs_equals(RStrs a, RStrs b) {
 static inline bool r_strs_startswith_strs(RStrs s, RStrs prefix) {
 	const size_t n = r_strs_len (prefix);
 	return n == 0 || (r_strs_len (s) >= n && !memcmp (s.a, prefix.a, n));
-}
-
-static inline bool r_strs_endswith_strs(RStrs s, RStrs suffix) {
-	const size_t n = r_strs_len (suffix);
-	const size_t l = r_strs_len (s);
-	return n == 0 || (l >= n && !memcmp (s.b - n, suffix.a, n));
 }
 
 /* C-string convenience wrappers — strlen folds to a constant for literals */
@@ -176,34 +97,12 @@ static inline bool r_strs_startswith(RStrs s, const char *prefix) {
 	return r_strs_startswith_strs (s, r_strs_from (prefix));
 }
 
-static inline bool r_strs_endswith(RStrs s, const char *suffix) {
-	return r_strs_endswith_strs (s, r_strs_from (suffix));
-}
-
-/* Comparison (out of line) */
-R_API int r_strs_cmp(RStrs a, RStrs b);
-R_API int r_strs_cmpi(RStrs a, RStrs b);
-
 /* Search (out of line primitives) */
 R_API const char *r_strs_find_strs(RStrs s, RStrs needle);
 R_API const char *r_strs_findc(RStrs s, char c);
-R_API const char *r_strs_rfindc(RStrs s, char c);
 R_API const char *r_strs_find_any(RStrs s, const char *set);
 
-static inline const char *r_strs_find(RStrs s, const char *needle) {
-	return r_strs_find_strs (s, r_strs_from (needle));
-}
-
-static inline bool r_strs_contains(RStrs s, const char *needle) {
-	return r_strs_find (s, needle) != NULL;
-}
-
-static inline bool r_strs_contains_char(RStrs s, char c) {
-	return r_strs_findc (s, c) != NULL;
-}
-
-/* Char-set trim, skip, tokenization (out of line) */
-R_API void r_strs_trim_chars(RStrs *s, const char *set);
+/* Char-set skip, tokenization (out of line) */
 R_API void r_strs_skip_chars(RStrs *s, const char *set);
 R_API RStrs r_strs_take_ident(RStrs *s);
 R_API bool r_strs_next_token(RStrs *s, const char *seps, RStrs *out);
@@ -247,13 +146,8 @@ static inline bool r_strs_split_strs(RStrs s, RStrs sep, RStrs *head, RStrs *tai
 	return true;
 }
 
-static inline bool r_strs_split_str(RStrs s, const char *sep, RStrs *head, RStrs *tail) {
-	return r_strs_split_strs (s, r_strs_from (sep), head, tail);
-}
-
 /* Conversion (out of line) */
 R_API char *r_strs_tostring(RStrs s);
-R_API size_t r_strs_ncopy(char *dst, size_t dstsize, RStrs s);
 R_API ut64 r_strs_tonum(RStrs s);
 R_API st64 r_strs_tosnum(RStrs s, bool *ok);
